+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Javascript SDK Client version was being sent in dispatched events. Changed it to send React SDK client version.
+
 ## [1.0.0] - September 27th, 2019
 
 Initial release